[arcane,impl] Déplace 'ItemGroupSynchronizeInfo' et 'VariableSyncInfo' dans 'DataSynchronizeInfo.h'.
